+++ Adjust/ADJLinkResolution.m
@@ -8,32 +8,25 @@
 
 #import "ADJLinkResolution.h"
 #import "ADJUtil.h"
-
 static NSUInteger kMaxRecursions = 10;
 
+// forward declaration for private methods
 @interface ADJLinkResolution (Private)
-
 + (BOOL)isTerminalUrlWithHost:(nullable NSString *)urlHost;
-
++ (nullable NSURL *)convertUrlToHttps:(nullable NSURL *)url;
++ (nullable NSString *)locationHeaderFromResponse:(nonnull NSHTTPURLResponse *)response;
++ (nonnull NSURLSession *)linkResolutionSession;
++ (void)requestAndResolveUrl:(nonnull NSURL *)url
+                   recursion:(NSUInteger)recursionNumber
+                    callback:(nonnull void (^)(NSURL *_Nullable resolvedLink))callback;
 @end
 
+// delegate to handle redirects
 @interface ADJLinkResolutionDelegate : NSObject<NSURLSessionTaskDelegate>
-
 + (nonnull ADJLinkResolutionDelegate *)sharedInstance;
-
-+ (nullable NSURL *)convertUrlToHttps:(nullable NSURL *)url;
-
-+ (NSURLRequest *)replaceUrlWithRequest:(NSURLRequest *)request
-                           urlToReplace:(nonnull NSURL *)urlToReplace;
-
 @end
 
 @implementation ADJLinkResolutionDelegate
-
-- (nonnull instancetype)init {
-    self = [super init];
-    return self;
-}
 
 + (nonnull ADJLinkResolutionDelegate *)sharedInstance {
     static ADJLinkResolutionDelegate *sharedInstance = nil;
@@ -48,40 +41,8 @@ static NSUInteger kMaxRecursions = 10;
                      willPerformHTTPRedirection:(NSHTTPURLResponse *)response
                                      newRequest:(NSURLRequest *)request
                               completionHandler:(void (^)(NSURLRequest * _Nullable))completionHandler {
-    // if we're already at a terminal host (adjust.com / adj.st / go.link),
-    // stop auto-following to preserve the terminal URL (avoid jumping to App Store links)
-    if ([ADJLinkResolution isTerminalUrlWithHost:response.URL.host]) {
-        completionHandler(nil);
-        return;
-    }
-
-    NSURL *_Nullable convertedUrl = [ADJLinkResolutionDelegate convertUrlToHttps:request.URL];
-
-    if (request.URL != nil && convertedUrl != nil && ![request.URL isEqual:convertedUrl]) {
-        completionHandler([ADJLinkResolutionDelegate replaceUrlWithRequest:request
-                                                              urlToReplace:convertedUrl]);
-    } else {
-        completionHandler(request);
-    }
-}
-
-+ (nullable NSURL *)convertUrlToHttps:(nullable NSURL *)url {
-    if (url == nil) {
-        return nil;
-    }
-    if (![url.absoluteString hasPrefix:@"http:"]) {
-        return url;
-    }
-
-    NSString *_Nonnull urlStringWithoutPrefix = [url.absoluteString substringFromIndex:5];
-    return [NSURL URLWithString:[NSString stringWithFormat:@"https:%@", urlStringWithoutPrefix]];
-}
-
-+ (NSURLRequest *)replaceUrlWithRequest:(NSURLRequest *)request
-                           urlToReplace:(nonnull NSURL *)urlToReplace {
-    NSMutableURLRequest *mutableRequest = [request mutableCopy];
-    [mutableRequest setURL:urlToReplace];
-    return [mutableRequest copy];
+    // prevent automatic redirects so we can resolve manually
+    completionHandler(nil);
 }
 
 @end
@@ -91,132 +52,176 @@ static NSUInteger kMaxRecursions = 10;
 + (void)resolveLinkWithUrl:(nonnull NSURL *)url
      resolveUrlSuffixArray:(nullable NSArray<NSString *> *)resolveUrlSuffixArray
                   callback:(nonnull void (^)(NSURL *_Nullable resolvedLink))callback {
-    if (callback == nil) {
+    // ignore if callback is not provided
+    if (!callback) {
         return;
     }
-    if (url == nil || url.host == nil || url.host.length == 0) {
+
+    // early returns for invalid URLs
+    if (!url || !url.host || url.host.length == 0) {
         [ADJUtil launchInMainThread:^{
             callback(url);
         }];
         return;
     }
 
-    // if suffix array is provided and URL doesn't match, return URL unchanged
-    if (![ADJLinkResolution urlMatchesSuffixWithHost:url.host
-                                         suffixArray:resolveUrlSuffixArray]) {
+    // only resolve when suffix array is provided and matches
+    if (!resolveUrlSuffixArray || ![self urlMatchesSuffixWithHost:url.host
+                                                      suffixArray:resolveUrlSuffixArray]) {
         [ADJUtil launchInMainThread:^{
             callback(url);
         }];
         return;
     }
 
-    ADJLinkResolutionDelegate *_Nonnull linkResolutionDelegate = [ADJLinkResolutionDelegate sharedInstance];
+    // convert HTTP to HTTPS for initial URL
+    NSURL *httpsUrl = [self convertUrlToHttps:url];
 
-    // reuse shared session for better performance
+    [self requestAndResolveUrl:httpsUrl
+                     recursion:0
+                       callback:callback];
+}
+
++ (nullable NSURL *)convertUrlToHttps:(nullable NSURL *)url {
+    if (!url || ![url.absoluteString hasPrefix:@"http:"]) {
+        return url;
+    }
+    return [NSURL URLWithString:[url.absoluteString stringByReplacingOccurrencesOfString:@"http:"
+                                                                              withString:@"https:"
+                                                                                 options:0
+                                                                                   range:NSMakeRange(0, 5)]];
+}
+
++ (nullable NSString *)locationHeaderFromResponse:(nonnull NSHTTPURLResponse *)response {
+    NSDictionary *headers = response.allHeaderFields;
+    for (id key in headers) {
+        if ([key isKindOfClass:[NSString class]] &&
+            [(NSString *)key caseInsensitiveCompare:@"Location"] == NSOrderedSame) {
+            id value = headers[key];
+            return [value isKindOfClass:[NSString class]] ? value : nil;
+        }
+    }
+    return nil;
+}
+
++ (nonnull NSURLSession *)linkResolutionSession {
     static NSURLSession *sharedSession = nil;
     static dispatch_once_t sessionOnceToken;
     dispatch_once(&sessionOnceToken, ^{
-        sharedSession =
-        [NSURLSession sessionWithConfiguration:NSURLSessionConfiguration.defaultSessionConfiguration
-                                      delegate:linkResolutionDelegate
-                                 delegateQueue:nil];
-    });
+        NSURLSessionConfiguration *config = [NSURLSessionConfiguration ephemeralSessionConfiguration];
+        config.timeoutIntervalForRequest = 3.0;
+        config.timeoutIntervalForResource = 8.0;
+        config.requestCachePolicy = NSURLRequestReloadIgnoringLocalCacheData;
+        config.URLCache = nil;
 
-    NSURL *_Nullable httpsUrl = [ADJLinkResolutionDelegate convertUrlToHttps:url];
-    NSURLSessionDataTask *task = [sharedSession dataTaskWithURL:httpsUrl
-                                              completionHandler:^(NSData * _Nullable data,
-                                                                  NSURLResponse * _Nullable response,
-                                                                  NSError * _Nullable error) {
-        // bootstrap the recursion of resolving the link
-        [ADJLinkResolution resolveLinkWithResponseUrl:response != nil ? response.URL : nil
-                                          previousUrl:httpsUrl
-                                      recursionNumber:0
-                                              session:sharedSession
-                                             callback:callback];
-    }];
-    [task resume];
+        sharedSession = [NSURLSession sessionWithConfiguration:config
+                                                      delegate:[ADJLinkResolutionDelegate sharedInstance]
+                                                 delegateQueue:nil];
+    });
+    return sharedSession;
 }
 
-+ (void)resolveLinkWithResponseUrl:(nullable NSURL *)responseUrl
-                       previousUrl:(nullable NSURL *)previousUrl
-                   recursionNumber:(NSUInteger)recursionNumber
-                           session:(nonnull NSURLSession *)session
-                          callback:(nonnull void (^)(NSURL *_Nullable resolvedLink))callback {
-    // return (possible nil) previous url when the current one does not exist
-    if (responseUrl == nil) {
-        [ADJUtil launchInMainThread:^{
-            callback(previousUrl);
-        }];
-        return;
-    }
-    // stop recursion when URL stops changing (prevents infinite loops)
-    if (previousUrl != nil && [responseUrl isEqual:previousUrl]) {
-        [ADJUtil launchInMainThread:^{
-            callback(responseUrl);
-        }];
-        return;
-    }
-    // return found url with expected host (Adjust terminal domains)
-    // these are domains where we stop to avoid redirecting to App Store
-    if ([ADJLinkResolution isTerminalUrlWithHost:responseUrl.host]) {
-        [ADJUtil launchInMainThread:^{
-            callback(responseUrl);
-        }];
-        return;
-    }
-    // return current url when it reached the max number of recursive tries
-    if (recursionNumber >= kMaxRecursions) {
-        [ADJUtil launchInMainThread:^{
-            callback(responseUrl);
-        }];
-        return;
-    }
++ (void)requestAndResolveUrl:(nonnull NSURL *)url
+                   recursion:(NSUInteger)recursionNumber
+                    callback:(nonnull void (^)(NSURL *_Nullable resolvedLink))callback {
+    NSMutableURLRequest *request = [NSMutableURLRequest requestWithURL:url];
+    request.HTTPMethod = @"GET";
 
-    // when found a non expected url host, use it to recursively resolve the link
-    NSURLSessionDataTask *task = [session dataTaskWithURL:responseUrl
-                                        completionHandler:^(NSData * _Nullable data,
-                                                            NSURLResponse * _Nullable response,
-                                                            NSError * _Nullable error) {
-        [ADJLinkResolution resolveLinkWithResponseUrl:response != nil ? response.URL : nil
-                                          previousUrl:responseUrl
-                                      recursionNumber:(recursionNumber + 1)
-                                              session:session
-                                             callback:callback];
+    NSURLSession *session = [self linkResolutionSession];
+    __block NSURLSessionDataTask *task =
+    [session dataTaskWithRequest:request
+               completionHandler:^(NSData * _Nullable data,
+                                   NSURLResponse * _Nullable response,
+                                   NSError * _Nullable error) {
+        NSURL *finalUrl = response.URL ?: (error ? task.currentRequest.URL : url) ?: url;
+        if (error != nil) {
+            [ADJUtil launchInMainThread:^{
+                callback(finalUrl);
+            }];
+            return;
+        }
+
+        NSHTTPURLResponse *httpResponse = [response isKindOfClass:[NSHTTPURLResponse class]]
+            ? (NSHTTPURLResponse *)response
+            : nil;
+        if (httpResponse == nil) {
+            [ADJUtil launchInMainThread:^{
+                callback(finalUrl);
+            }];
+            return;
+        }
+
+        NSString *location = [self locationHeaderFromResponse:httpResponse];
+        if (location.length > 0) {
+            NSURL *redirectUrl = [NSURL URLWithString:location relativeToURL:finalUrl];
+            redirectUrl = [self convertUrlToHttps:redirectUrl.absoluteURL];
+
+            if (redirectUrl == nil) {
+                [ADJUtil launchInMainThread:^{
+                    callback(finalUrl);
+                }];
+                return;
+            }
+
+            if ([self isTerminalUrlWithHost:redirectUrl.host]) {
+                [ADJUtil launchInMainThread:^{
+                    callback(redirectUrl);
+                }];
+                return;
+            }
+
+            if (recursionNumber + 1 > kMaxRecursions) {
+                [ADJUtil launchInMainThread:^{
+                    callback(redirectUrl);
+                }];
+                return;
+            }
+
+            // stop recursion when redirect doesn't change the URL
+            if ([redirectUrl isEqual:finalUrl]) {
+                [ADJUtil launchInMainThread:^{
+                    callback(finalUrl);
+                }];
+                return;
+            }
+
+            [self requestAndResolveUrl:redirectUrl
+                             recursion:recursionNumber + 1
+                              callback:callback];
+            return;
+        }
+
+        [ADJUtil launchInMainThread:^{
+            callback(finalUrl);
+        }];
     }];
     [task resume];
 }
 
 + (BOOL)isTerminalUrlWithHost:(nullable NSString *)urlHost {
-    if (urlHost == nil) {
+    if (!urlHost) {
         return NO;
     }
 
-    // check hardcoded Adjust terminal domains
-    // these are domains where we stop recursion to avoid redirecting to App Store
-    NSArray<NSString *> *_Nonnull terminalUrlHostSuffixArray = @[@"adjust.com",
-                                                                 @"adj.st",
-                                                                 @"go.link",
-                                                                 @"adjust.cn",
-                                                                 @"adjust.net.in",
-                                                                 @"adjust.world",
-                                                                 @"adjust.io"];
+    NSArray<NSString *> *terminalSuffixes =
+    @[@"adjust.com", @"adj.st", @"go.link", @"adjust.cn", @"adjust.net.in", @"adjust.world", @"adjust.io"];
+    for (NSString *suffix in terminalSuffixes) {
+        if ([urlHost hasSuffix:suffix]) {
+            return YES;
+        }
+    }
 
-    return [ADJLinkResolution urlMatchesSuffixWithHost:urlHost
-                                           suffixArray:terminalUrlHostSuffixArray];
+    return NO;
 }
 
 + (BOOL)urlMatchesSuffixWithHost:(nullable NSString *)urlHost
                      suffixArray:(nullable NSArray<NSString *> *)suffixArray {
-    if (urlHost == nil) {
+    if (!urlHost || !suffixArray) {
         return NO;
     }
 
-    if (suffixArray == nil) {
-        return NO;
-    }
-
-    for (NSString *_Nonnull expectedHostSuffix in suffixArray) {
-        if ([urlHost hasSuffix:expectedHostSuffix]) {
+    for (NSString *suffix in suffixArray) {
+        if ([urlHost hasSuffix:suffix]) {
             return YES;
         }
     }